part1/authentication/permissions.py
from rest_framework import permissions

from authentication.models import User


class VacancyCreatePermission(permissions.BasePermission):
    message = 'Adding vacancies for non hr user not allowed.'

    def has_permission(self, request, view):
        if request.user.role != User.HR:
            return False
        return True